+++ lib/bloc/theme/theme_state.dart
@@ -2,24 +2,20 @@ part of 'theme_bloc.dart';
 
 @immutable
 abstract class ThemeState {
-  final bool prefsAreLoaded;
-
-  const ThemeState({required this.prefsAreLoaded});
+  const ThemeState();
 
   AppTheme get theme;
 }
 
 class LightThemeState extends ThemeState {
-  const LightThemeState({bool prefsAreLoaded = false})
-      : super(prefsAreLoaded: prefsAreLoaded);
+  const LightThemeState();
 
   @override
   AppTheme get theme => LightTheme();
 }
 
 class DarkThemeState extends ThemeState {
-  const DarkThemeState({bool prefsAreLoaded = false})
-      : super(prefsAreLoaded: prefsAreLoaded);
+  const DarkThemeState();
 
   @override
   AppTheme get theme => DarkTheme();
