+++ lib/view/components/search/language_selector.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:get_it/get_it.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../../models/themes/theme.dart';
@@ -11,45 +12,40 @@ class LanguageSelector extends StatefulWidget {
 }
 
 class _LanguageSelectorState extends State<LanguageSelector> {
-  late final SharedPreferences prefs;
+  final SharedPreferences prefs = GetIt.instance.get<SharedPreferences>();
   late List<bool> isSelected;
 
   @override
   void initState() {
     super.initState();
-    isSelected = [false, false, false];
-
-    SharedPreferences.getInstance().then((prefs) {
-      this.prefs = prefs;
-      setState(() {
-        isSelected = _getSelectedStatus() ?? isSelected;
-      });
-    });
+    isSelected = _getSelectedStatus() ?? [false, false, false];
   }
 
-  void _updateSelectedStatus() {
-    prefs.setStringList(
-      'languageSelectorStatus',
-      isSelected.map((b) => b ? '1' : '0').toList(),
+  Future<void> _updateSelectedStatus() async => prefs.setStringList(
+        'languageSelectorStatus',
+        isSelected.map((b) => b ? '1' : '0').toList(),
+      );
+
+  List<bool>? _getSelectedStatus() => prefs
+      .getStringList('languageSelectorStatus')
+      ?.map((s) => s == '1')
+      .toList();
+
+  Widget _languageOption(String language) => 
+    Container(
+      padding: const EdgeInsets.symmetric(vertical: 10.0, horizontal: 20.0),
+      child: Center(child: Text(language)),
     );
-  }
-
-  List<bool>? _getSelectedStatus() {
-    return prefs
-        .getStringList('languageSelectorStatus')
-        ?.map((s) => s == '1')
-        .toList();
-  }
 
   @override
   Widget build(BuildContext context) {
     return ToggleButtons(
       selectedColor: AppTheme.jishoGreen.background,
       isSelected: isSelected,
-      children: const <Widget>[
-        _LanguageOption('Auto'),
-        _LanguageOption('日本語'),
-        _LanguageOption('English')
+      children: <Widget>[
+        _languageOption('Auto'),
+        _languageOption('日本語'),
+        _languageOption('English')
       ],
       onPressed: (buttonIndex) {
         setState(() {
@@ -59,20 +55,6 @@ class _LanguageSelectorState extends State<LanguageSelector> {
           _updateSelectedStatus();
         });
       },
-    );
-  }
-}
-
-class _LanguageOption extends StatelessWidget {
-  final String language;
-
-  const _LanguageOption(this.language);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(vertical: 10.0, horizontal: 20.0),
-      child: Center(child: Text(language)),
     );
   }
 }
