+++ lib/view/components/search/search_bar.dart
@@ -22,9 +22,7 @@ class SearchBar extends StatelessWidget {
               ),
             ),
           ),
-          const SizedBox(
-            height: 10.0,
-          ),
+          const SizedBox(height: 10.0),
           const LanguageSelector()
         ],
       ),
