+++ pubspec.yaml
@@ -6,13 +6,13 @@ environment:
   sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  animated_size_and_fade: ^2.0.0
+  animated_size_and_fade: ^3.0.0
   confirm_dialog: ^1.0.0
   division: ^0.9.0
   flutter:
     sdk: flutter
-  flutter_bloc: ^7.0.1
-  flutter_slidable: ^0.6.0
+  flutter_bloc: ^8.0.0
+  flutter_slidable: ^1.1.0
   get_it: ^7.2.0
   mdi: ^5.0.0-nullsafety.0
   path: ^1.8.0
